No RT access from "OscilSmp/" & "FMSmp/" anymore
This removes the ports "OscilSmp/" and "FMSmp/" dispatching of any
OscilGen ports. They are now forwarded in any case.

While the ports are removed, the objects OscilGn and FmGn remain in
ADnoteVoiceParam. This is probably just to make XML load/save easier,
or to keep consistency (MW also did not store PAD note's non RT params,
it only kept pointers).

This fix will still work for automation: If a host sends non-RT
messages to Master, they will be forwarded to MiddleWare.